Impl Level to Struct

+++ programs/zai/src/lib.rs
@@ -13,12 +13,12 @@ pub mod zai {
     use crate::instructions::{create_player as cp, change_default_class as cdc, modify_player_xp as mpxp};
 
     // Wrapper for create_player
-    pub fn create_player(ctx: Context<CreatePlayer>, active_class: u64, active_weapon: u64) -> Result<()> {
+    pub fn create_player(ctx: Context<CreatePlayer>, active_class: u8, active_weapon: u8) -> Result<()> {
         cp(ctx, active_class, active_weapon)
     }
 
     // Wrapper for change_default_class
-    pub fn change_default_class(ctx: Context<ChangeDefaultClass>, new_class: u64) -> Result<()> {
+    pub fn change_default_class(ctx: Context<ChangeDefaultClass>, new_class: u8) -> Result<()> {
         cdc(ctx, new_class)
     }
